Initial version for showing full class body
