+++ plugin/clang/utils.py
@@ -6,7 +6,6 @@ Attributes:
 import platform
 import logging
 import subprocess
-import html
 
 from os import path
 
@@ -151,103 +150,3 @@ class ClangUtils:
         # if we haven't found anything there is nothing to return
         log.error("no libclang found at all")
         return None
-
-    @staticmethod
-    def get_text_by_extent(extent):
-        """Load lines of code in range, pointed by extent.
-
-        Args:
-            extent (Cursor.extent): Ranges of source file.
-        """
-        if extent.start.file.name != extent.end.file.name:
-            return None
-
-        with open(extent.start.file.name, 'r') as f:
-            lines = f.readlines()
-            return "".join(lines[extent.start.line - 1:extent.end.line])
-
-    @staticmethod
-    def htmlize_text(text):
-        """Convert plain text to HTML like representation.
-
-        Args:
-            text: Plain text to convert.
-        """
-        result = html.escape(text)
-        result = result.replace("\\\n", "<br>")
-        result = result.replace("\n", "<br>")
-        result = result.replace("\t", 4 * "&nbsp;")
-        result = result.replace(" ", "&nbsp;")
-
-        # sublime does not recognize &quot;
-        result = result.replace("&quot;", "&nbsp;")
-        return result
-
-    @staticmethod
-    def build_objc_message_info_details(cursor):
-        """Provide information about cursor to Objective C message expression.
-
-        Builds detailed information about cursor when cursor is
-        a CursorKind.OBJC_MESSAGE_EXPR. OBJC_MESSAGE_EXPR cursors
-        behave very differently from other C/C++ cursors in that:
-        - The return type we want to show in the tooltip
-          is stored in the original 'cursor.type' from the cursor the user is
-          hovering over; in C/C++ we only used 'cursor.referenced' but nothing
-          else from the original cursor.
-        - 'cursor.referenced' is still important, as it holds the name and args
-          of the method being called in the message. But
-          'cursor.referenced.spelling' comes in a different format then what
-          For example, if we have this method declaration for 'bar':
-            @interface Foo
-              -(void)bar:(BOOL)b1 boolParam2:(BOOL):b2
-            @end
-          And later, we hover over the text calling bar():
-            Foo* foo = [[Foo alloc] init];
-            [foo bar:YES boolParam2:NO]; // <- Hover over 'bar' here
-          Then we would see:
-            cursor.kind = CursorKind.OBJC_INSTANCE_METHOD_DECL
-            cursor.type.spelling = 'void'
-            cursor.referenced.kind: CursorKind.OBJC_INSTANCE_METHOD_DECL
-            cursor.referenced.spelling = 'bar:boolParam2:'
-            cursor.referenced.arguments[0].type.spelling = 'BOOL'
-            cursor.referenced.arguments[0].spelling = 'b1'
-            cursor.referenced.arguments[1].spelling = 'BOOL'
-            cursor.referenced.arguments[1].spelling = 'b2'
-          Our goal is to make the tooltip match the method declaration:
-            'void bar:(BOOL)b1 boolParam2:(BOOL):b2'
-        - Objective C methods also don't need to worry about static/const
-
-        Args:
-            cursor (Cursor): Current cursor.
-        """
-        result = ""
-        return_type = cursor.type
-        result += ClangUtils.link_from_location(
-            ClangUtils.location_from_type(return_type),
-            html.escape(return_type.spelling))
-
-        result += ' '
-
-        method_cursor = cursor.referenced
-        method_and_params = method_cursor.spelling.split(':')
-        method_name = method_and_params[0]
-        if method_cursor.location:
-            result += ClangUtils.link_from_location(
-                method_cursor.location,
-                html.escape(method_name))
-        else:
-            result += html.escape(method_cursor.spelling)
-
-        method_params_index = 1
-        for arg in method_cursor.get_arguments():
-            result += ":(" + arg.type.spelling + ")"
-            if arg.spelling:
-                result += arg.spelling + " "
-            result += method_and_params[method_params_index]
-            method_params_index += 1
-
-        if method_cursor.brief_comment:
-            result += "<br><br><b>"
-            result += method_cursor.brief_comment + "</b>"
-
-        return result
